FEATURE: Add ESBuild bundling system
As the Rollup.js bundling and build system cannot cope with the native '.node' binary files in the VSCode extensions, it is needed to add an alternative bundling system which can do it.

ESBuild is a high speed, high performance bundler written in GO, quickly gaining a popularity between TypeScript and JavaScript developers. The support for '.node' modules is very unofficial, but for the current needs of PathPilot Robot VSCode extension acceptable.

The development of the extension in Visual Studio Code can now use both Rollup based and ESBuild based debugging.